Refactor to use middleware code from weaveworks/common
