+++ scripts/build.sh
@@ -23,7 +23,6 @@ CODE_DIR=$(cd $SCRIPT_DIR/..; pwd)
 echo $CODE_DIR
 
 cp -r $CODE_DIR/cmd/ $CODE_DIR/docker/payment/cmd/
-cp -r $CODE_DIR/middleware/ $CODE_DIR/docker/payment/middleware/
 cp $CODE_DIR/*.go $CODE_DIR/docker/payment/
 mkdir $CODE_DIR/docker/payment/vendor && cp $CODE_DIR/vendor/manifest $CODE_DIR/docker/payment/vendor/
